fix: Setup the dynamic import so it is separate from other packages (#113)
* fix: Setup the dynamic import so it is separate from other packages

The default setup for dynamic import was such that the libraries would
overlap with other libraries having the same names, which caused version
mismatches when using the dynamic import version.  This change makes the
path separate.  As well, it allows a named load for the library, which
means the library can actually be used natively in the browser again,
fixing issues with deployment standalone.

* Add access to a few internals so that external features work